+++ src/modules/basic_integrators/implicit_PicardBDE_integrator.f90
@@ -42,6 +42,9 @@ module implicit_PicardBDE_integrator_class
         integer(ik) :: minNonlinIters = 5 !! Number of nonlinear iterations below which the number of substeps gets reduced
 
         integer(ik) :: maxRestarts = 3 !! Maximum number of consecutive solver restart attempts before critical failure is announced
+
+        integer(ik) :: hardMaxRestarts = 10 !! Maximum number of consecutive solver restart attempts before critical failure is announced regardless of whether consolidation happened or not
+
         integer(ik) :: restartCount = 0 !! Counter for consecutive number of solver restars 
 
         integer(ik) :: consolidationInterval = 50 !! How many integration calls before currentNumSubsteps is again reduced to 1