+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## v1.1.0, 2023-09-26
+## v1.1.0, 2023-11-21
 
 - Solver and integrator improvements
 - Debug mode bug fixes
@@ -10,6 +10,8 @@
 ### New Features
 
 - The BDE integrator internal controller now attempts to reduce the number of substeps back to 1 after every 50 successful integrations
+- The BDE integrator now also has a hard maximum of attempted restarts regardless of whether consolidation happens. this avoids crashes due to allocating too much memory for the timestep size buffer
+- The BDE integrator now prints the approximate convergence bottleneck variable (feature not completely reliable)
 - Command line PETSc support for setting up the KSP solver object has been implemented. The command line ksp_type takes precedent over the integrator options. This now allows for command line customization of the KSP object. Note that tolerances are still set from config options. 
 - New derivation type nDLinInterpDerivation - linear interpolation on n-dimensional data
 - ModelboundCRMData can now have an associated non-default electron species ID (useful for multiple flux tube models). Note that Janev transitions still have hardcoded values.